test(recordings): keep projection coverage owner-local
